⏯️ Move continue to its own crate
