+++ src/commands/mod.rs
@@ -1,4 +1,5 @@
 pub mod auth;
+pub mod cont;
 pub mod list;
 pub mod running;
 pub mod stop;
