+++ src/constants.rs
@@ -1,0 +1,1 @@
+pub const CLIENT_NAME: &str = "github.com/heytherewill/toggl-cli";
